+++ src/python_file_execution_request.h
@@ -2,24 +2,26 @@
 
 #include <memory>
 #include <string>
-#include <json/value.h>
 
-namespace PythonRuntime::pythonFileExecution {
+#include "json/value.h"
+
+namespace PythonRuntime::PythonFileExecution {
 
 struct PythonFileExecutionRequest {
-    std::string file_execution_path = "";
-    std::string python_library_path = "";
-    bool isDefaultLib = true;
+  std::string file_execution_path = "";
+  std::string python_library_path = "";
+  bool isDefaultLib = true;
 };
 
-inline PythonFileExecutionRequest fromJson(std::shared_ptr<Json::Value> json_body) {
-    PythonFileExecutionRequest request;
+inline PythonFileExecutionRequest FromJson(std::shared_ptr<Json::Value> json_body) {
+  PythonFileExecutionRequest request;
 
-    if (json_body) {
-        request.file_execution_path = json_body->get("file_execution_path", "").asString();
-        request.python_library_path = json_body->get("python_library_path", "").asString();
-    }
-    return request;
+  if (json_body) {
+    request.file_execution_path = json_body->get("file_execution_path", "").asString();
+    request.python_library_path = json_body->get("python_library_path", "").asString();
+  }
+
+  return request;
 }
 
-} // namespace PythonRuntime::pythonFileExecution
+} // namespace PythonRuntime::PythonFileExecution
